+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.6.2
+* Added `overflow` method to `StyleClass`.
+* [Breaking] `angled` property inside the `elevation` method is renamed to `angle` and is now given in a circular value. 0.0 equals down. [null] equals directly under the widget.
+
 ## 0.6.1
 * [Fix] `ripple` and `opacity` now works with `add` method.
 * Other minor improvements and fixes